Simplify the OracleSlabCache interface. (#4)
The next() method no longer requires the allocate, ready and swap functions, as
we switch to a std::list in the internal implementation that avoids the need to
fiddle around with the Slab object itself after construction. 

This also provides cleaner enforcement of the maximum number of slabs. We extend
some of the tests to check that no additional allocations are performed.