Support subset-aware oracle caches in CustomChunkedMatrix. (#6)
This enables more efficient iteration through the CustomChunkedMatrix if the
Chunk_ class supports extraction of a subset of the primary dimension. 

This change required the addition of a indexed primary extract() method to the
Chunk_ contract. We implemented this in the simple wrappers as a demonstration.

We also modified the simple wrappers so that the positions of the extracted
subset in the output of each extract() call is the same as if we had extracted
the entire chunk, so the former can be a drop-in replacement for the latter.

A consequence of this change to the contract is that we need a separate method
to extract a single primary element without requiring a full slab allocation.
This was also implemented in the simple wrappers.