Simplified code for the oracle slab cache classes.  (#11)
This is made possible by the new tatami::Oracle interface, which allows
multiple passes through a finite stream of predictions. Thus, we can iterate
through all predictions to predict future usage, without having to allocate
memory to hold the same predictions for "present-day" retrieval.

There is no longer any need for a maximum number of predictions per cycle, as
we can just go all the way to the end (or until the cache is filled). This also
eliminates the need for a free store of unused slabs in the cache, as we can
guarantee that all cached slabs are used in each cycle, or we're at the end.

The definition of populate() is also much simplified. Callers just have to 
operate on a single vector of to-be-populated slabs, rather than using the first
vector to index into the second argument to obtain the slab.

Finally, we implement a minor optimization by skipping the unordered_map
look-up if consecutive predictions are on the same slab.