Delete copy construction/assignment for cache classes.
This is necessary as the cache classes hold iterators/pointers to their
own internal members; any copy would have incorrect references to the
original, and I can't be bothered to go through and update them on copy.

Move construction/assignment is still okay, as the targets of the
iterators and pointers are moved into the new object.