Replaced TypicalSlabCacheWorkspace with the simpler SlabCacheStats. (#14)
We now require each extractor subclass to explicitly state the cache that it's
using, which is easier to understand. To align with this new approach, we
refactor CustomChunkedMatrix extractors into multiple subclasses based on the
caching policy. While this adds more boilerplate code, it actually improves
readability by making sure the code for each cache type is separated.