Use contiguous memory pool for slabs in the custom chunked matrices. (#15)
This performs one big allocation during Extractor construction, avoiding repeated
allocations when creating new slabs. The idea is to be more accurately follow
the cache size memory limit; otherwise, lots of small allocations may be
rounded up such that the real memory usage exceeds the limit. There might also
be some performance benefit as we don't hit the allocator so frequently.

This requires some modification of the sparse chunk contract to accept pointers
rather than vectors. At this point, given that we're modifying the contract, we
also allow the values/indices of non-zero elements to be optional on output
from the sparse chunk's extract methods. This allows chunk developers to skip
storing the values/indices if they were not requested during extraction.