Base Extractor classes don't need explicit destructors.
We never call the destructor through a pointer to these base classes, so
there isn't any need to construct a vtable for them. The only vtable
that is required is for the interface's destructor, which looks for and
calls the derived class's destructor, which then automatically calls the
destructor for all of its base classes.