Removed the requirement for an index_type for sparse chunks.
This is because each chunk doesn't know how long the non-target
dimension actually is. Thus, an integer type that can store the indices
of a single chunk may not be large enough to store the indices of a
longer slab. The only way to be sure is to store the indices as the
Index_ type of the parent matrix, which eliminates the need for u to
consider the chunk's index type during caching.