Shift row-related template parameters to regular run-time arguments. (#16)
This aligns with the policy in tatami core, i.e., template parameters should
only be used for types and not to excessively optimize at compile-time. The row
access choice is now pushed all the way to the Chunk_'s extract() method; this
is probably more intuitive than templating and shouldn't degrade performance
in realistic applications where chunk decompression is not trivial.

Also fixed the sparse matrix test so that multiple cache sizes are used.
Removed some unnecessary methods from the chunk coordinator.