Optionally report the re-used slabs in OracularSlabCache::next().
This allows users to perform some housekeeping on the re-used slabs before or
after populating the new slabs, in a similar manner to the variable cache. It
is intended for cases where the slabs share memory and we want to defragment
the memory pool to create a contiguous buffer for third-party libraries to fill
the memory for multiple slabs at once.