Test that the oracular caches start a new caching cycle correctly.
Specifically, a new cycle (i.e., a round of populating the cache with multiple
slabs) should only be performed when the cache is full and a new slab is
encountered that is not in the current cache; otherwise, this would imply an
unnecessary cache miss. This test just ensures that the caching is optimal.

Some work is required to implement this test for the OracularVariableSlabCache,
due to the differences between the actual/estimated sizes for each slab.