Switch to an abstract interface for chunk data in the CustomChunkedMatrix classes. (#19)
This allows developers to choose between run- and compile-time polymorphism for
the chunk data, to avoid having to instantiate multiple classes for different
chunking/compression schemes. It also documents the expected interface in more
natural way that can be enforced by the compiler.

The interface itself is a Manager class that performs the extraction given the
grid coordinates of the chunk. This gives implementations more flexibility as
to how chunks are stored, rather than forcing everyone to be in a vector. 
We create separate Manager interfaces for the dense and sparse cases.

The Manager interface doesn't have a (compile-time) choice between full and
subsetted extraction on the target dimension. This simplifies implementations
as they don't have to consider two sets of extract() methods. The choice is now
instead done at run-time, which reduces binary bloat when compiling with a
variety of compression schemes that may or may not benefit from subsetting.