Help devirtualization by adding the new_workspace_exact() method.
Implementations of Manager interfaces can implement this method to return a
pointer to the concrete Workspace class (instead of a pointer to the base
class). Then, if an application wants to use a concrete Manager in their
realized ChunkedMatrix class, all Extractors will automatically contain
pointers to the concrete Workspace to avoid virtual dispatch.

We do this as it seems (from some godbolt testing) that neither clang
nor GCC can devirtualize automatically in this case. Note that this method
is strictly optional - it is only required for a Manager implementation if
an application uses that concrete subclass directly as the Manager_.