Check for integer overflow when computing products and resizing containers. (#20)
These checks use the sanisizer library and should mostly be no-ops, but they
provides some protection against weird size_type definitions.

We template the (maximum) number of slabs so that we can check for overflow
inside the constructors of our various classes.

We cap the default cache sizes to avoid compiler warnings on overflow. 

We also switch the prediction indices to tatami::PredictionIndex to further 
reduce the arbitrary reliance on std::size_t.