Cut down on excessive use of templated constructors.
The original motivation was to check for integer overflow when converting from
the user-supplied type to the class member. However, this should really be the
responsibility of the caller. We already expect the caller to handle protection
for standard library calls and virtual methods where templating is impossible,
so there is no need to make an exception for constructor calls.

This choice simplifies the constructors and makes them more self-documenting,
as the class developer can express their intended type when an argument is
copied directly to a class member. We also ensure that CTAD remains possible,
which would otherwise fail if the constructor has separate templated params.